Use explicit AWS URL for EvaporateJS
Makes the upload service take an explicit S3 URL to be used by EvaporateJS. This prevents Evaporate from constructing the URL based on other parameters, and allows for supporting S3 endpoints that Evaporate does not, such as dualstack endpoints. When a aws_url is used, the cloudfront parameter must be set, so this change hard codes that value.